Linked List Methods
This file only has the methods for the linked list. The only methods that need to get worked on are the Convert(), Remove() and Replace(). The Convert() method is the bonus method we get to choose for the assignment. Another issue is that I can't figure out how to print the user's data as a string. It only prints --> Assignment3.User

If you wanna give it a try to figure it out, be my guest,